Remove special handling of name attribute. Fixes #278
The name attribute gets special treatment in the codebase, in that
it's not simply yet another attribute but stored in it's own field.

Thus, every callside dealing with attributes needs to be aware of
this special case and explicitly handle the name attribute. It's
easy to see that this can lead to bugs. For example, querying for
the name attribute is currently broken due the querying code not
being aware of the special case [1].

Discussions with experienced bazel developers came to the conclusion
that there is no need (anymore) to treat the name attribute specially
and thus we decided it's best to remove the special treatment and
handle the name attribute as any other attribute.

This change removes the handling of name attributes and also adds a test
case to verify that bug [1] is fixed.

[1] https://github.com/bazelbuild/bazel/issues/278

--
PiperOrigin-RevId: 147446345
MOS_MIGRATED_REVID=147446345