Bazel client, Windows: impl. ForEachDirectoryEntry
This is the last function we needed from
file_posix for MSYS, so now we can remove that
from the compilation.

Fixes https://github.com/bazelbuild/bazel/issues/2386
The problem originally was that I used CloseHandle
to close the HANDLE, instead of using FindClose,
so we were holding on to open directory HANDLEs,
so we couldn't rename the installation directory.

See https://github.com/bazelbuild/bazel/issues/2107

--
Change-Id: If6c1b3c99cf386d82e829dbee2323e6bce5f6b46
Reviewed-on: https://cr.bazel.build/8952
PiperOrigin-RevId: 147734165
MOS_MIGRATED_REVID=147734165