Tune naming of Android data binding metadata files.
This puts Bazel in line with the paths actually produced by the data
binding resource processor.

This difference matters when the library name doesn't match the package
name (the build fails without this change).

I started writing a shell integration test to check this end-to-end. But
staging all the data binding support libs and their deps is a mess. Too much
so, IMO.

--
PiperOrigin-RevId: 147868538
MOS_MIGRATED_REVID=147868538