Skylark documentation pipeline automatically generates documentation for all Skylark modules.
With this change, we no longer have to manually register modules whose generation should be generated. Manual registration turned out to be prone to errors since engineers didn't know about it or simply forgot to register their modules. As a result, we had ~20 modules that featured nice documentation texts, but that were not being picked up by the documentation pipeline.

--
PiperOrigin-RevId: 148136776
MOS_MIGRATED_REVID=148136776