Add a flag to enable desugaring java.* classes by rewriting package names
java.* and sun.* classes cannot be desugared directly because there are hard coded restrictions in the JVM that prevent using loading or using reflection on them.  Add a --core_library flag that rewrites package names to have a __desugar__ prefix when reading them, and strips the prefix when writing them back out.

--
PiperOrigin-RevId: 148273386
MOS_MIGRATED_REVID=148273386