Remove AbtractAction#getInputsForExtraAction() (the third time).
This time, if the action discovers inputs, the extra action is made to depend on the outputs of the action so that by the time the extra action runs, the inputs of the original action are discovered. This avoids us having to think about the state the shadowed action may keep.

Yes, actions should not keep state, but they do. Such is life.

--
PiperOrigin-RevId: 148627715
MOS_MIGRATED_REVID=148627715