sandbox should create regular empty files, not symlink to /dev/null.
Adds a test based on our Python rules that makes sure that this actually
fixes the issue. Thanks to @duggelz for the suggestion. It seems like
our Python rules are the only place that actually provides an
EmptyFilesSupplier to Runfiles, so there's probably no simpler way to
test this behavior in an integration test.

Fix #1458. Fix #2394.

--
PiperOrigin-RevId: 148656193
MOS_MIGRATED_REVID=148656193