BEP: Make TestSummary provide the overall status
Bazel does an involved computation of the overall outcome of a test,
taking into considerations multiple attempts to detect flaky tests.
To allow a consumer of the BEP to have the same view of the outcome
of the test as bazel does, report the overall status of a test in the
TestSummary event.

Note: as proto-to-proto dependencies are not yet available, we
have to provide the status by duplicating the status enum.

--
Change-Id: I413ce7b72c54ff3cbfe0733219ba0881ad5d064e
Reviewed-on: https://cr.bazel.build/9129
PiperOrigin-RevId: 149268441
MOS_MIGRATED_REVID=149268441