Add a source_jars variable to OutputJar.
Having a correlation between an output jar and a source jar is not enough.
There may be situations when an output jar is generated from more source jars,
not just one. We need this flexibility especially in Skylark for the java
sandwich, when the user can compile multiple source jars.

--
PiperOrigin-RevId: 149510534
MOS_MIGRATED_REVID=149510534