Make gold autodetection more robust on macos
Some macos clang versions don't fail when setting -fuse-ld=gold, which makes
cc_configure.bzl think gold linker is available, when it is actually not.

The sideeffect of this cl is that we will not autodetect gold when only a very
old (year 2010 and older) is present. I assume this is fine.

Fixes #2502

--
PiperOrigin-RevId: 149521872
MOS_MIGRATED_REVID=149521872