A partial, manual rollback of commit 4eb06ce15042c962e1c262cbf20a251d6c8d6a41.
This became necessary because extra actions for C++ compile actions require .h files, but the compiler only returns the .pcm files in the .d file for headers that it reads from the .pcm file. This is not a problem for correctness because the .pcm files depend on the headers, but that doesn't help the extra actions that would then only get the .pcm files.

--
PiperOrigin-RevId: 150052839
MOS_MIGRATED_REVID=150052839