Change the UUID of actions consuming the set of build info keys.
This should have been part of commit 34fb97e5a7a93f47be50562b128278a367f757f6 which in itself causes problems if the Blaze version used changes because the language-specific header actions are then erroneously cached.

Ideally, we'd add the set of build info headers into the key, but that's not available in computeKey(). The next best thing would be to reify them as artifacts, but that's a little too intrusive for the urgency of the issue.

--
PiperOrigin-RevId: 150299218
MOS_MIGRATED_REVID=150299218