BEP: Indicate caching status in TestResult
For a test result indicate if it was cached locally by the tool. Note
that this bit will not be set for tests cached by the remote execution.
Also clarify the meaning of the timing; it need not be set for locally
cached test results, but it may (and in this case reporting the timing
of the initial invocation of the cached test).

--
Change-Id: Ic83c4a58c1a85f6f4da82138999eaa545d8f60a6
Reviewed-on: https://cr.bazel.build/9376
PiperOrigin-RevId: 150442579
MOS_MIGRATED_REVID=150442579