Define the ConfigFeatureFlagConfiguration fragment.
This will eventually be used to store the values of user-defined
configuration flags, once such a rule is added. Because this rule
does not exist yet, this fragment is currently not part of the
rule class provider. This will have to be done when that rule
is turned on.

--
PiperOrigin-RevId: 150638292
MOS_MIGRATED_REVID=150638292