Enable SkylarkCallable methods to be used on ClassObjects.
Previously, ClassObjects' SkylarkCallable struct fields worked
as expected (with values of the struct taking precedence), but
calling methods annotated with SkylarkCallable did not work at
all; methods were treated as if they were not present in the struct.

This adds some tests for the various ways of looking up fields
and methods, and rearranges the logic in FuncallExpression to
allow for callable methods on SkylarkClassObject's descendants.

--
PiperOrigin-RevId: 150876181
MOS_MIGRATED_REVID=150876181