sandbox: No longer require a temporary "sandbox" directory.
By removing the --sandbox_block_path feature in an earlier change and
taking advantage of the fact that in a mount namespace we can actually
"remount" mount points to be read-only without bind mounting them to
some other place beforehand, this is no longer necessary. The code
becomes much simpler due to this, for example we no longer need to
chroot.

--
PiperOrigin-RevId: 151111360
MOS_MIGRATED_REVID=151111360