Keep compile-from-source instructions distribution agnostic
The instructions on how to compile bazel from source are particularly
useful for platforms where no ready-made binaries are provided for.
As such, we should not assume a debian-package--based system. It is
OK to mention what needs to be done on those systems, but only as
an example after stating the goal to achive (have OpenJKD 8 installed).

PiperOrigin-RevId: 151571121